fix: 총 금액이 int 범위를 초과하는 경우에 대한 오류 수정

+++ app/src/main/java/woowacourse/shopping/domain/model/CartItem.kt
@@ -5,6 +5,6 @@ data class CartItem(
     val product: Product,
     val amount: Int,
 ) {
-    val totalPrice: Int
-        get() = product.price.value * amount
+    val totalPrice: Long
+        get() = (product.price.value * amount).toLong()
 }
